Add SocketAI to FrontEnd GUI
Can now test SocketAI via the front end provided by MicroRTS. However,
note that upon selecting the AI, it will automatically try to connect,
to the server. This cannot be changed without modifying the SocketAI
as it tries to perform the connection in its constructor. Also, note
the default language choice is `LANGUAGE_XML`.